refactor: remove unused code and update code structure
